wrds_datastream/requirements.txt
wrds
pandas
python-decouple
